додав авторизацію firebase
1) додав авторизацію
2) створив user в localstorege по якому будемо відображати кнопку або 'увійти' або 'вийти'
3) в хедері ДжС дописав що потрібно зробити
4) в хедер ДжС додав також змінну юсер якак буде приходити якщо авторизувався
