+++ package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "axios": "^1.5.0",
+    "firebase": "^10.4.0",
     "modern-normalize": "^1.1.0",
     "notiflix": "^3.2.6",
     "tui-pagination": "^3.4.1"
@@ -28,7 +29,8 @@
     "buffer": "^6.0.3",
     "parcel": "^2.6.0",
     "parcel-reporter-clean-dist": "^1.0.4",
-    "posthtml-include": "^1.7.4"
+    "posthtml-include": "^1.7.4",
+    "process": "^0.11.10"
   },
   "browserslist": "> 0.5%, last 2 versions, not dead"
 }
