--- src/js/login/fire-base.js
@@ -1,12 +1,0 @@
-const firebaseConfig = {
-    apiKey: "AIzaSyBcETIWRzdvvBXoViQftnTvFFLpHkebOhw",
-    authDomain: "book-bc5e2.firebaseapp.com",
-    projectId: "book-bc5e2",
-    storageBucket: "book-bc5e2.appspot.com",
-    messagingSenderId: "258455649332",
-    appId: "YOUR_APP_ID"
-  };
-
-  
-  
-
